Add test git repository helper to replace fixtures.
Fixtures for this kind of testing can be problematic as we don't know for sure which version of git
will be running during the tests. Creating a temp directory and running git commands inside of it
ensures that the installed version of git will understand how to interact with the data, and avoids
the odd .git directory swapping altogether.