don't override GIT_SSH_COMMAND from environment
The behavior of duplicate environment variables was previously
undefined, so in go1.9 the exec package was updated to specify that the
last instance of a variable wins, allowing for the expected behavior
when overriding a variable.

Rather than appending the GIT_SSH_COMMAND to the Command environment,
which would override previous settings, check the process environment
first, and append our options to any existing command.