Set atime and mtime of extracted tar files
As the atime and mtime of files is not currently set, they default to
the time they were extracted instead of the times in the original files.

We must take care to set the atime and mtime of the directories after
all files have been extracted, since the addition of a file to a
directory changes its mtime.