Use a new SSH key and obfuscate it a tiny bit.
GitHub scans and disables SSH keys where the private half is found in
raw form in code some place. This is great! But, go-getter has this SSH
key used to test cloning private GitHub repositories. It's safe since
it's a read-only, repo-scoped deploy key. To avoid the scanner, we'll
try a simple string split / concatenation.