HttpGetter: Add ability to set headers
This adds the ability to set headers that will be sent out on every
request of a particular HttpGetter. This is useful in situations where
query parameters are not suitable, such as when headers are explicitly
expected, or when one wants to move information off the query string
that would be at risk of possibly being exposed in logs or error
messages.

Fixes #71.