Revert "merge tmpFile and tempFile usages"
This reverts commit 7607f9ca4cfadd17dac92c56d009da72407b14ab.

The old test tempFile function is not general purpose and sets up a
static file within a directory created by the tempDir helper. This
functionality needs to be preserved, so the correct plan of action would
be to preserve this functionality while moving the function name to
another, logical name such as tempTestFile.