parse url on windows: fix bug where and absolute path got prefixed with "./"
RFC 3986 §4.2
A path segment that contains a colon character (e.g., "this:that") cannot be used as the first segment of a relative-path reference, as it would be mistaken for a scheme name. Such a segment must be preceded by a dot-segment (e.g., "./this:that") to make a relative-path reference.

Read "net/url".URL.String() for more info