Adds support for Symlinks in all Tar decompressors
This is useful when using Nomad's Artefact mechanisms which internally
uses `hashicorp/go-getter` to unpack Tar archives that are built from
container root file systems (*which usually contain symlinks for runtime
libraries*).

Fixes #60 and supercedes the very old PR #37

(We use a variant of this patch in our production systems and we need to
get this into upstream as soon as conveniently possible)

**NB:** This uses the
[securejoin](https://github.com/cyphar/filepath-securejoin) library to
prevent path escaping and other possible security vulnerabilities that
come with concatenating paths.