go get cloud.google.com/go@v0.45.1
The primary motivation of this is that this old version of
cloud.google.com/go depends on an old version of go.opencensus.io which
in turn depends on git.apache.org/thrift.git, which is an old mirror of
the Thrift repository that is no longer canonical and is currently down.

We don't need thrift in our dependency graph at all here, and we only
had it in the first place because go.opencensus.io was depending on it.
Newer versions of that library don't include a thrift client in the main
library and so we don't inherit that transitive dependency anymore.

This also cleans up some accumulated checksums that we no longer need from
the go.sum file.