[1 of 6] getter.Detect(...): extract function handleDetected(...)
Once the Detect(...) function finds a detector for a source string it
combines the obtained result string with other data bits to produce the
result returned to the caller. That processing is here extracted into a
new handleDetected(...) function, which will be called from an
additional context in an upcoming commit.

The new handleDetected(...) function lives in a new file:

    detect_common.go

The move emphasizes the function's slightly wider use now by both the
Detect(...) and (soon-to-be-introduced) CtxDetect(...) dispatch
functions. With the introduction of CtxDetect, the logic in
handleDetected is no longer used exclusively by Detect.

This changeset provides an implementation modification, but no
behavioral change.

Signed-off-by: Alan D. Salewski <ads@salewski.email>