Let shallow clone work with any ref
Currently the shallow clone only works if ref is either not set, or set
to the default branch.  This is becuase go-getter does not set the ref
to checkout during a shallow clone so Git assumes the default branch,
and then when go-getter tries to checkout a ref other than the default
branch it fails because the shallow clone did not fetch that ref.

To fix this, I have used `--branch` to explicitly set the ref to fetch
during a clone so that when a shallow clone is done go-getter will clone
the ref given and not the default branch.  This will also make clones of
non-default branches just a tad bit more optimized since the clone will
not waste time checking out a branch it ultimately will not use.