[3 of 6] Add full set of stock CtxDetector implementations
...and uncomment the 'ContextualDetectors' defult list of them in
CtxDetector.

The CtxDetector implementations satisfy the CtxDetector interface
for all of the built-in detectors (bitbucket, github, gitlab, file,
git, GCS, and S3), but do not (yet) take advantage of the the
additional contextual information available. These implementations
all just pass-through a subset of their arguments to the existing
Detector impl. analogue.

These are intended to make it comfortable to swap-in use of the new
"Contractual Detector" without having to change much code, and also
provide a place to hang enhancement code that can take advantage of
the contextual information made available by the CtxDetect dispatch
function.

Signed-off-by: Alan D. Salewski <ads@salewski.email>