[4 of 6] Add unit test for full set of stock CtxDetector impls.
...except for GitCtxDetector, which will be getting special treatment
since it is the module behind the motivation for this series of changes.

The unit tests committed here are all just minimal adaptations of the
existing detector unit tests to satisfy the CtxDetect method (which
takes three more params than Detect). These tests basically test the
"pass through" behavior of their respective implementations, but that's
useful to demonstrate that they do not interfere with the traditional
behaviors.

Signed-off-by: Alan D. Salewski <ads@salewski.email>