Add role option for S3 URLs
Add support for specifying an IAM role that is assumed before making S3
requests. Beforehand this was only possible when using profiles which
required local files for configuration.

Now it is possible to specify such an IAM role in the URL, thereby
supporting settings where a IAM role needs to be assumed for data
access (e.g. cross-account).

Example of usage:
    s3::bucket.s3.amazonaws.com/file?aws_iam_role=some-role-arn