Fix support for getting single files from GitHub
This currently fails for GitHub (and possibly other Git repositories)
because there is an assumption in the "client.go" file that the "subDir"
segment is always a directory.

However, for this URL:

https://github.com/owner/repo/subdir/file.txt

The "subDir" variable would actually be a file ("subdir/file.txt") and
hence the client should use "copyFile" rather than "copyDir" when
copying the "subDir" to the final destination.

This change fixes that to check whether the "subDir" variable is a file
or a directory, and appropriately uses "copyDir" or "copyFile". It also
returns the single file in the "GetResult" object, if indeed a single
file was requested.

Tests attached.