fix(getter): race condition when SetClient is called
this is a non-breaking change. this fixes an issue when Clients are created
and called concurrently. clients called SetClient on each of the
Getters. When no getters were provided, it modified the clients
of the default getters defined in a map. Since the map was mutated
from multiple goroutines, it creates a race condition 'concurrent map writes'.
The issue is mitigated by creating a new copy of Getters for every new client
created. SetClient now modifies its own copy of the Getter, not the default Getters