gcs/v2: bump cloud.google.com/go/storage to 1.35.0
The storage library that gcs depends on transitively imports grpc in
version 1.50.0, which is vulnerable to GHSA-m425-mq94-257g.

While this is a server-side vulnerability, therefore the package is not
directly vulnerable (nor its clients), this dependency still causes
advisories to be produced against this package, so we update those now.