Handle failed decompressed files
When untarring a tarball, attempt to create files initially with the
umask file permissions and delete the file if it fails to copy
correctly. This prevents a paritally written file from remaining on disk
with permissions that didn't include the umask.

Adds a unit test and a invalid tarball (created by truncating a large
file generated from /dev/urandom), which fails to decompress but errors
partially written.